hack/graph.html: Add arrowheads
Based on Dino Fancellu's [1] (Released under the GNU General Public
License, version 3) and [2].  This uses marker-end to place the marker
at the end of the path, orient=auto to align the marker with the
tangent to the path at that end, and then refX to shift the arrowhead
back so that the point touches the edge of the node circle (instead of
touching the center of the node circle and thus being masked by that
circle).  A drawback to this approach is that the linear tangent at
the end of the path will diverge from curving paths, leaving
arrowheads that are not centered on, and in some cases miss
completely, the line to which they belong.  Would be nice to fix that,
ideally with something like "position this marker markRadius pixels
short of the end of the path", but I'm not conversant enough in SVG to
know how to do that.

The more-targetted element removal allows us to continue to inject new
dynamic elements like nodes and edges without blowing away the static
arrowhead marker element.

[1]: http://bl.ocks.org/fancellu/2c782394602a93921faff74e594d1bb1
[2]: https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/marker-end