hack/graph.html: Make channel configurable
There's no current API for "give me all the channels Cincinnati knows
about", so hard-code stable-4.2 as the default.  I'm also injecting
any value given via the user's query parameter, so they can select a
different channel (e.g. candidate-4.3).  If they pick an invalid
channel, they get no error, but will see an empty graph.  In the
future, we might want to explicitly show something like "no nodes for
${channel}", but I'm punting on that for now.

For each graph we fetch, I'm also walking the returned nodes and
including any channels to which those nodes also belong.  That's
useful for going to earlier 4.y or for discovering different stability
channels in your current 4.y.  For example, the stable-4.2 graph will
turn up candidate-4.2, stable-4.1, prerelease-4.1, etc.  It will not
turn up stable-4.3, because none of the nodes in 4.3 will ever be in
stable-4.2.  In the future, we might have a stable channel that is not
pinned to a particular 4.y, and if/when that happens, the default
should be changed to use that channel.

While new channels are injected in alphabetical order, I am punting on
sorting them into existing entries.  That would be nice, but is beyond
the level of polish I currently have time for.